requirements.txt
# requirements.txt
# Developer: G A RAZA
# Premium ChatGPT-like Telegram Bot Dependencies

python-telegram-bot==20.7
openai==1.3.7
langdetect==1.0.9
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.1.0
reportlab==4.0.7
markdown==3.5.1
beautifulsoup4==4.12.2
asyncio==3.4.3